Add state syncer implementation

+++ src/state/include/state/interfaces/i_state_syncer.h
@@ -5,8 +5,6 @@
 
 #pragma once
 
-#include "state/interfaces/i_command_giver.h"
-#include "state/interfaces/i_command_taker.h"
 #include "state/player_state.h"
 #include "state/utilities.h"
 
@@ -54,7 +52,7 @@ class STATE_EXPORT IStateSyncer {
      *
      * @param game_over Signal that the game is over
      */
-    virtual std::array<int64_t, 2> getScores(bool game_over) = 0;
+    virtual std::array<size_t, 2> getScores(bool game_over) = 0;
 };
 
 } // namespace state
